pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Modular System for Unstructured-to-Structured Data Transformation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.7",
    "tiktoken>=0.6.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "nltk>=3.8.1",
    "jsonschema>=4.21.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.0",
    "asyncio>=3.4.3",
    "dspy>=2.6.27",
    "openai>=1.97.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["assignment"]

[tool.uv.sources]
assignment = { path = "." }
